chore(PocketIC): refactor progress threads in PocketIC
